Update reading on a cadence
We really do not need to update the reading _every_ time we go through
the loop for the socket.  This is a sump hole, not an electron beam.
This should save some CPU cycles for other things.

Also tossed a gc.collect in there.  Hopefully that sorts the issue where
the wESP32 has been crashing out with a "OSError: [Errno 113]
ECONNABORTED" after some random period of time.