Tried/Failed to do a neat moving average
Turns out that the main loop blocks on the socket accept.  So the loop
to update the moving average only runs when a new request comes in.
Will need to thread the webserver out to make that work.  A task for
another time.  Going to see if the gc.collect() at least keeps this
stable for a while.